refactor gametrakr
