+++ web/src/components/auth/index.ts
@@ -1,2 +1,3 @@
-export * from "./sign-in-dialog"
-export * from "./sign-up-dialog"
+export * from "./auth-dialog"
+export * from "./sign-in-form"
+export * from "./sign-up-form"
